feat(match2): parse participant into arrays
