+++ components/match2/commons/match_group_display_helper.lua
@@ -169,32 +169,6 @@ function DisplayHelper.MapScore(score, opponentIndex, resultType, walkover, winn
 	return score and tostring(score) or ''
 end
 
----@param allParticipants table<string, table>
----@param opponentIndex integer
----@return table[]
-function DisplayHelper.getParticipantsOfOpponent(allParticipants, opponentIndex)
-	local prefix = opponentIndex .. '_'
-	local function indexFromKey(key)
-		if String.startsWith(key, prefix) then
-			return tonumber(string.sub(key, #prefix + 1))
-		else
-			return nil
-		end
-	end
-	local participantsOfOpponent = Array.extractValues(Table.mapArguments(
-		allParticipants,
-		indexFromKey,
-		function (key, index)
-			if Logic.isEmpty(allParticipants[key]) then
-				return nil
-			end
-			return Table.merge({playerId = index}, allParticipants[key])
-		end,
-		true
-	))
-	return Array.sortBy(participantsOfOpponent, Operator.property('playerId'))
-end
-
 --[[
 Display component showing the detailed summary of a match. The component will
 appear as a popup from the Matchlist and Bracket components. This is a
