+++ components/match2/wikis/valorant/match_summary.lua
@@ -359,25 +359,18 @@ end
 function CustomMatchSummary._createMap(game)
 	local row = MatchSummary.Row()
 
-	local team1Agents, team2Agents
-
-	if not Table.isEmpty(game.participants) then
-		team1Agents = Agents():setLeft()
-		team2Agents = Agents():setRight()
-
-		for _, playerStats in ipairs(DisplayHelper.getParticipantsOfOpponent(game.participants, 1)) do
-			team1Agents:add(playerStats.agent)
-		end
-		for _, playerStats in ipairs(DisplayHelper.getParticipantsOfOpponent(game.participants, 2)) do
-			team2Agents:add(playerStats.agent)
-		end
+	local team1Agents = Agents():setLeft()
+	local team2Agents = Agents():setRight()
+	for _, playerStats in ipairs((game.opponents[1] or {}).players or {}) do
+		team1Agents:add(playerStats.agent)
+	end
+	for _, playerStats in ipairs((game.opponents[2] or {}).players or {}) do
+		team2Agents:add(playerStats.agent)
 	end
 
-	local score1, score2
-
 	local extradata = game.extradata or {}
-	score1 = Score():setLeft()
-	score2 = Score():setRight()
+	local score1 = Score():setLeft()
+	local score2 = Score():setRight()
 
 	score1:setMapScore(DisplayHelper.MapScore(game.scores[1], 1, game.resultType, game.walkover, game.winner))
